// Aerocode/enum.ts
export enum tipoAeronave{
    COMERCIAL = "Comercial",
    MILITAR = "Militar"
}

export enum tipoPeca{
    NACIONAL = "Nacional",
    IMPORTADA = "Importada"
}

export enum statusPeca{
    PRODUÇÃO = "Em produção",
    TRANSPORTE = "Em transporte", 
    PRONTA = "Pronta para uso"
}

export enum producao{
    PENDENTE = "Pendente",
    ANDAMENTO = "Em andamento", 
    CONCLUIDO = "Concluído"
}

export enum hierarquia{
    ADMINISTRADOR = "Administrador",
    ENGENHEIRO = "Engenheiro",
    OPERADOR = "Operador"
}

export enum testes{
    ELETRICOS = "Elétrico",
    HIDRAULICOS = "Hidraulico",
    AERODINAMICO = "Aerodinamico"
}

export enum result{
    APROVADO = "Aprovado",
    REPROVADO = "Reprovado"
}